requirements/base.txt
Django==4.2.3
Flake8-pyproject==1.2.3
Pillow==10.0.0
PyJWT==2.7.0
bandit==1.7.5
black==23.7.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.3
factory-boy==3.2.1
flake8==6.0.0
freezegun==1.2.2
gunicorn==20.1.0
isort==5.12.0
mixer==7.2.2
pip==23.1.2
pipdeptree==2.9.5
psycopg2==2.9.6
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-xdist==3.3.1
pytest==7.4.0
python-decouple==3.8
requests==2.31.0
wheel==0.40.0
